test(unified-item): split unifiedItem tests into separate files by module
Split the original unifiedItem test suite into individual test files for childOperations, conversionUtils, migrationUtils, treeUtils, validateItemHierarchy, and unifiedItemSchema. Improves test organization and maintainability.
